+++ tests/testthat/test-tar_make.R
@@ -91,3 +91,21 @@ tar_test("tar_make() handles callr errors", {
   )
   expect_null(NULL)
 })
+
+tar_test("priorities apply to tar_make() (#437)", {
+  skip_on_cran()
+  tar_script(
+    list(
+      tar_target(x1, 1, priority = 0),
+      tar_target(y1, 1, priority = 0.5),
+      tar_target(z1, 1, priority = 1),
+      tar_target(x2, x1, priority = 0),
+      tar_target(y2, y1, priority = 0.5),
+      tar_target(z2, z1, priority = 1)
+    )
+  )
+  tar_make(callr_function = NULL)
+  out <- tar_progress()$name
+  exp <- c("z1", "z2", "y1", "y2", "x1", "x2")
+  expect_equal(out, exp)
+})