+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Enhancements
 
+* Enable priorities in `tar_make()`.
 * Optionally show priorities in `tar_manifest()`.
 * Show the priority in the print method of stem and pattern targets.
 * Throw informative errors if the secondary arguments to `pattern = slice()` or `pattern = sample()` are invalid.